requirements.txt
certifi==2020.6.20
chardet==3.0.4
Click==7.0
cycler==0.10.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-MQTT==1.0.5
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
greenlet==0.4.15
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
monotonic==1.5
mysqlclient==1.4.4
numpy==1.20.3
paho-mqtt==1.4.0
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
scikit-learn==0.24.2
scipy==1.6.3
six==1.12.0
SQLAlchemy==1.3.10
threadpoolctl==2.1.0
typing==3.7.4.1
urllib3==1.25.10
Werkzeug==0.16.0
